+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.6.21 (2023/06/08)
+
+-   ## Rename Style File
+-   ## Add `FlexWrap` property
+-   ## Add breakpoint object
+
 ## 0.6.20 (2023/06/06)
 
 -   ## Create RvSidebarContainer and RvSidebarHeader